test(contract): move dispute tests to integration suite
